version bump v0.10.0-beta
- New initialization process. React.renderComponent no longer takes
Astarisx application arguments. Initialization occurs in the `ui`
component in `componentWillMount` using `this.initializeAppContext`
which takes the necssary arguments plus rest (...) arguments to be
passed to the ControllerViewModel `dataContextWillInitialize`.
- `this.intializeDataContext` now accepts either no args, '*', '_*',
dataContext names as strings or an Object with dataContext names as
keys with the value being the args to be passed to each dataContext
during initialization.